+++ myuw/dao/instructor.py
@@ -1,10 +1,8 @@
 import logging
-from restclients_core.exceptions import DataFailureException
 from myuw.models import SeenInstructor
 from uw_sws.section import get_last_section_by_instructor_and_terms
 from myuw.dao.pws import get_person_of_current_user
-from myuw.dao.term import get_prev_num_terms,\
-    get_term_before, get_previous_quarter
+from myuw.dao.term import get_term_before, get_previous_quarter
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +22,7 @@ def is_instructor(request):
 
     person = get_person_of_current_user(request)
     user_netid = person.uwnetid
-    if is_seen_instructor(user_netid):
+    if SeenInstructor.is_seen_instructor(user_netid):
         request.myuw_is_instructor = True
         return True
 
@@ -39,22 +37,8 @@ def is_instructor(request):
         quarter = section.term.quarter
         year = section.term.year
         try:
-            add_seen_instructor(user_netid, year, quarter)
+            SeenInstructor.add_seen_instructor(user_netid, year, quarter)
         except Exception as ex:
             logger.error("add_seen_instructor(%s, %s, %s) ==> %s",
                          user_netid, year, quarter, ex)
     return request.myuw_is_instructor
-
-
-def is_seen_instructor(uwnetid):
-    return SeenInstructor.objects.filter(uwnetid=uwnetid).exists()
-
-
-def add_seen_instructor(netid, year, quarter):
-    SeenInstructor.objects.update_or_create(uwnetid=netid,
-                                            quarter=quarter,
-                                            year=year)
-
-
-def remove_seen_instructors_yrs_before(year):
-    SeenInstructor.objects.filter(year__lt=year).delete()
